% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaffke_CI.R
\name{gaffke_CI}
\alias{gaffke_CI}
\title{Gaffke bound as described in Learned-Miller and Thomas https://arxiv.org/pdf/1905.06208.pdf}
\usage{
gaffke_CI(x, alpha = 0.05, B = 10000, side = "upper")
}
\arguments{
\item{x}{A vector of iid random samples from some distribution F with mean mu}

\item{alpha}{The level; the output is a (1-alpha) confidence bound}

\item{B}{the number of Monte Carlo iterations to estimate the interval}

\item{side}{the side of the confidence interval, either "upper" or "lower"}
}
\value{
An upper or lower level alpha confidence bound on the population mean mu
}
\description{
Gaffke bound as described in Learned-Miller and Thomas https://arxiv.org/pdf/1905.06208.pdf
}
