% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sample_gaffke.R
\name{two_sample_gaffke}
\alias{two_sample_gaffke}
\title{Two-sample test of mean dominance}
\usage{
two_sample_gaffke(
  sample_1,
  sample_2,
  B = 1000,
  method = "fisher",
  bounds = c(0, 1)
)
}
\arguments{
\item{sample_1}{a numeric vector, a random sample from a population with unknown mean \eqn{\mu_1}}

\item{sample_2}{a numeric vector, a random sample from a second population with unknown mean \eqn{\mu_2}}

\item{B}{a positive integer, the number of Monte Carlo iterations to be used in running the test}

\item{method}{a string, the method of combining two one-sample tests into a two-sample test of equality
\itemize{
\item Sidak: using Sidak's correction, compute a 1-sqrt(1-alpha) upper and lower bound (respectively) and see if they overlap
\item Fisher: compute a combined p-value using Fisher's combining function of whether both means are equal to a particular mu_0, then maximize this over possible values of mu_0
\item Liptak: compute a combined p-value using Liptak's combining function
}}

\item{bounds}{a length-2 vector of the (shared) upper and lower bounds on both populations, [a,b]}
}
\value{
A P-value indicating for the hypothesis test H_0: \eqn{\mu_2 <= \mu_1}
}
\description{
A function to run a one-sided, two-sample nonparametric test based on the Gaffke (Learned-Miller and Thomas) test.
In particular, given IID samples from two distributions with means \eqn{\mu_1} and \eqn{\mu_2} the hypothesis H_0: \eqn{\mu_2 <= \mu_1} is tested.
The test assumes that the distributions share common bounds [a,b], where a is the smallest and b is the largest value that samples from the populations can take.
If the distribution is contained within these bounds, the test is valid (P-values are dominated by the uniform distribution) no matter the shape of the population distributions.
Standard two-sample tests (e.g., the t-test) are not valid over all bounded distributions in finite-samples, and may fail to control the significance level.
}
\examples{
two_sample_gaffke(sample_1 = rbeta(30, shape1 = 5, shape2 = 0.5), sample_2 = rbeta(30, shape1 = 100, shape2 = 10))
}
